T: move short type rendering tests to a separate file
